full rewrite of additionalBundlerResources-usage on mac
The first idea was too naive, so this needed to be reimplemented completely. This now gets worked around and refactored a lot. Additionally the gradle-plugin got some more options, especially the options `failOnError`, `onlyCustomBundlers` and `skipJNLP` got added. The later one mostly makes sense in combination with `failOnError`, because the JNLP-bundler behaves different. Another important change is the switch default of the `checkForAbsolutePaths` option, which will get removed on the next release (I do know at least one person, who will be happy about this decision).